first styling pass on setscreen and datasetitem

+++ src/components/DataSetItem.js
@@ -1,5 +1,13 @@
 import React, { Component } from 'react';
-import { View, Button } from 'react-native';
+import {
+  StyleSheet,
+  Text,
+  Alert,
+  View,
+  TouchableOpacity,
+  Button,
+} from 'react-native';
+import { colors } from './../constants/colors';
 
 class DataSetItem extends Component {
   state = {
@@ -16,8 +24,24 @@ class DataSetItem extends Component {
   }
 
   deleteSet() {
-    // TODO find out why delete opens button for new item
-    this.props.deleteItem(this.props.title);
+    Alert.alert(
+      `${this.props.title}`,
+      'Are you sure you want to delete this set?',
+      [
+        {
+          text: 'Cancel',
+          style: 'cancel',
+        },
+        {
+          text: 'Delete',
+          onPress: () => {
+            this.setState({ showButtons: false });
+            this.props.deleteItem(this.props.title);
+          },
+        },
+      ],
+      { cancelable: false },
+    );
   }
 
   render() {
@@ -28,17 +52,44 @@ class DataSetItem extends Component {
           onPress={() => this.onSetNamePress()}
         />
         {this.state.showButtons && (
-          <View>
-            <Button
-              title="Add Data Point"
+          <View style={styles.buttonsViewStyle}>
+            <TouchableOpacity
+              style={styles.buttonStyle}
               onPress={() => this.onAddDataPoint()}
-            />
-            <Button title="Delete Data Set" onPress={() => this.deleteSet()} />
+            >
+              <Text style={styles.buttonTextStyle}>View/Add Data</Text>
+            </TouchableOpacity>
+            <TouchableOpacity
+              style={[styles.buttonStyle, styles.deleteStyle]}
+              onPress={() => this.deleteSet()}
+            >
+              <Text style={styles.buttonTextStyle}>Delete Set</Text>
+            </TouchableOpacity>
           </View>
         )}
       </View>
     );
   }
 }
+
+const styles = StyleSheet.create({
+  buttonsViewStyle: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-around',
+  },
+  buttonStyle: {
+    width: 125,
+    backgroundColor: colors.buttonBackground,
+    borderRadius: 5,
+    overflow: 'hidden',
+  },
+  deleteStyle: {
+    backgroundColor: colors.delete,
+  },
+  buttonTextStyle: {
+    textAlign: 'center',
+  },
+});
 
 export default DataSetItem;
